Regional plot save

fig_region_save saves the regional plots to file.

Usage

fig_region_save(
  fig,
  file,
  ntraits = 1,
  interactive = FALSE,
  width = NULL,
  height = NULL,
  dpi = 500
)

Arguments

fig

a regional plot object

file

a character string of the file path to save the plot

ntraits

an integer value of the number of traits plotted (default: 1)

interactive

a logical value indicating whether the plot is interactive (default: FALSE)

width

a numeric value indicating the width of the plot in inches (default: NULL)

height

a numeric value indicating the height of the plot in inches (default: NULL)

dpi

the resolution of the plot (default: 500)

Author

James Staley jrstaley95@gmail.com

Examples

if (FALSE) { # \dontrun{
fig <- fig_region(
  data = geni.plots::geni_test_region$assoc,
  corr = geni.plots::geni_test_region$corr,
  build = 37
)
fig_region_save(fig, "test.png")
} # }