GENI plots test data for the stacked regional plot

An R object containing formatted results from genome-wide association studies (GWAS) of interleukin-6 levels (GCST90274815) and interleukin-6 receptor levels (GCST90088597) and linkage disequilibrium statistics from 1000 Genomes Phase 3 for the region 1:154301970-154551970.

Usage

geni_test_stack_region

Format

A list containing 2 objects with the information necessary for constructing an interactive stacked regional plot. Specifically, this includes the following information:

assoc a data.frame with genetic association results for interleukin-6 levels from GCST90274815 and for interleukin-6 receptor levels GCST90088597 for the region 1:154301970-154551970

corr a matrix of correlation statsitics from the European samples of 1000 Genomes Phase 3 for the region 1:154301970-154551970, this matrix has the same markers in the same order as assoc

Details

assoc is a data.frame with the following columns:

marker the genomic marker identifier (i.e. rsID or chromosome-position)

chr the chromosome for each genomic marker

pos the genomic position (GRCh37) for each genomic marker

pvalue_1 the association p-value for interleukin-6 levels for each genomic marker

pvalue_2 the association p-value for interleukin-6 receptor levels for each genomic marker

Author

James Staley jrstaley95@gmail.com

Wes Spiller

Examples

head(geni_test_stack_region$assoc)
#>        marker chr       pos pvalue_1 pvalue_2
#> 1   rs6691345   1 154302826 7.56e-06 6.95e-05
#> 2 rs115165870   1 154303228 8.76e-03 1.08e-11
#> 3  rs12727865   1 154303705 5.47e-02 2.09e-09
#> 4 rs147889978   1 154303783 1.21e-01 6.30e-22
#> 5   rs6702754   1 154303976 5.58e-02 2.09e-09
#> 6   rs1209264   1 154305010 4.01e-02 4.68e-04
head(geni_test_stack_region$corr)
#>             rs6691345 rs115165870 rs12727865 rs147889978 rs6702754 rs1209264
#> rs6691345      1.0000     -0.0514    -0.2433     -0.0477   -0.2433    0.1543
#> rs115165870   -0.0514      1.0000    -0.1444      0.0348   -0.1444    0.0528
#> rs12727865    -0.2433     -0.1444     1.0000      0.2230    1.0000    0.1941
#> rs147889978   -0.0477      0.0348     0.2230      1.0000    0.2230    0.0687
#> rs6702754     -0.2433     -0.1444     1.0000      0.2230    1.0000    0.1941
#> rs1209264      0.1543      0.0528     0.1941      0.0687    0.1941    1.0000
#>             rs12077870 rs6427575 rs28409746 rs2483710 rs11586379 rs1760796
#> rs6691345      -0.2433    1.0000    -0.3340    0.2148     0.5858    0.2148
#> rs115165870    -0.1444   -0.0514     0.1973    0.1335    -0.0883    0.1335
#> rs12727865      1.0000   -0.2433    -0.6323   -0.3621     0.3481   -0.3621
#> rs147889978     0.2230   -0.0477    -0.1296    0.1288     0.2723    0.1288
#> rs6702754       1.0000   -0.2433    -0.6323   -0.3621     0.3481   -0.3621
#> rs1209264       0.1941    0.1543     0.3378    0.5946     0.2111    0.5946
#>             rs150779199 rs34880256 rs2988721 rs9787014 rs34870927 rs6427576
#> rs6691345       -0.0286    -0.1331    0.2148   -0.1331    -0.3340    0.5858
#> rs115165870     -0.0037    -0.1195    0.1335   -0.1195     0.1973   -0.0883
#> rs12727865       0.1921     0.6132   -0.3621    0.6132    -0.6323    0.3481
#> rs147889978      0.7298    -0.1001    0.1288   -0.1001    -0.1296    0.2723
#> rs6702754        0.1921     0.6132   -0.3621    0.6132    -0.6323    0.3481
#> rs1209264        0.0665     0.1225    0.5946    0.1225     0.3378    0.2111
#>             rs2481064 rs1760795 rs767804 rs1352333 rs7515370 rs1626035
#> rs6691345      0.2148   -0.3340   1.0000    0.2148    0.5858    0.2148
#> rs115165870    0.1335    0.1973  -0.0514    0.1335   -0.0883    0.1335
#> rs12727865    -0.3621   -0.6323  -0.2433   -0.3621    0.3481   -0.3621
#> rs147889978    0.1288   -0.1296  -0.0477    0.1288    0.2723    0.1288
#> rs6702754     -0.3621   -0.6323  -0.2433   -0.3621    0.3481   -0.3621
#> rs1209264      0.5946    0.3378   0.1543    0.5946    0.2111    0.5946
#>             rs1760794 rs2481065 rs1194588 rs34968830 rs35463210 rs2010828
#> rs6691345     -0.3340    0.1543   -0.5819    -0.1350    -0.2142    1.0000
#> rs115165870    0.1973    0.0528    0.0887    -0.1204    -0.1333   -0.0514
#> rs12727865    -0.6323    0.1941   -0.3498     0.6083     0.3621   -0.2433
#> rs147889978   -0.1296    0.0687   -0.2736    -0.1010    -0.1286   -0.0477
#> rs6702754     -0.6323    0.1941   -0.3498     0.6083     0.3621   -0.2433
#> rs1209264      0.3378    1.0000   -0.2048     0.1243    -0.5918    0.1543
#>             rs749966 rs2274988 rs77993403 rs1194587 rs1194586 rs1979575
#> rs6691345    -0.1331   -0.2142    -0.1072   -0.3340   -0.3340   -0.1709
#> rs115165870  -0.1195   -0.1333    -0.0453    0.1973    0.1973   -0.0619
#> rs12727865    0.6132    0.3621    -0.1167   -0.6323   -0.6323    0.6380
#> rs147889978  -0.1001   -0.1286    -0.0436   -0.1296   -0.1296    0.3731
#> rs6702754     0.6132    0.3621    -0.1167   -0.6323   -0.6323    0.6380
#> rs1209264     0.1225   -0.5918     0.0580    0.3378    0.3378    0.1204
#>             rs1194585 rs11809740 rs1194584 rs148299382 rs1194583 rs2297607
#> rs6691345     -0.3340    -0.2142   -0.3361     -0.1503   -0.9953   -0.1969
#> rs115165870    0.1973    -0.1333    0.1969     -0.0507    0.0524   -0.1220
#> rs12727865    -0.6323     0.3621   -0.6317     -0.1933    0.2402    0.3081
#> rs147889978   -0.1296    -0.1286   -0.1306     -0.0667    0.0487   -0.1039
#> rs6702754     -0.6323     0.3621   -0.6317     -0.1933    0.2402    0.3081
#> rs1209264      0.3378    -0.5918    0.3332     -0.9894   -0.1460   -0.6123
#>             rs2297606 rs41310887 rs12729561 rs3811452 rs17699328 rs11589949
#> rs6691345     -0.2168    -0.0828    -0.1412   -0.1331     1.0000    -0.1709
#> rs115165870   -0.1474    -0.0714    -0.1121   -0.1195    -0.0514    -0.0619
#> rs12727865     0.4013    -0.1333     0.5870    0.6132    -0.2433     0.6380
#> rs147889978   -0.1294    -0.0086    -0.0924   -0.1001    -0.0477     0.3731
#> rs6702754      0.4013    -0.1333     0.5870    0.6132    -0.2433     0.6380
#> rs1209264     -0.4996     0.0662     0.1152    0.1225     0.1543     0.1204
#>             rs80282309 rs1194592 rs1552483 rs2483712 rs1194589 rs35735180
#> rs6691345      -0.0805   -0.3265    0.1598   -0.3216    0.1586    -0.1375
#> rs115165870    -0.0390    0.1952    0.0647    0.1960   -0.3229     0.3391
#> rs12727865      0.1486   -0.6220   -0.6244   -0.6201   -0.5207     0.4867
#> rs147889978    -0.0383   -0.1319   -0.3689   -0.1324   -0.3635     0.3801
#> rs6702754       0.1486   -0.6220   -0.6244   -0.6201   -0.5207     0.4867
#> rs1209264       0.0421    0.3131   -0.1080    0.3209   -0.1273     0.1245
#>             rs4040747 rs6688376 rs1889312 rs11576151 rs140907814 rs72633645
#> rs6691345      0.9321   -0.1981    0.4458     0.9035      0.1349     0.8672
#> rs115165870   -0.0425    0.2327    0.1816    -0.0471     -0.0114    -0.0540
#> rs12727865    -0.2410    0.7632    0.5231    -0.2183     -0.0591    -0.1871
#> rs147889978   -0.0385    0.2833    0.2299    -0.0431     -0.0098    -0.0500
#> rs6702754     -0.2410    0.7632    0.5231    -0.2183     -0.0591    -0.1871
#> rs1209264      0.1684    0.1941    0.2871     0.1755      0.0765     0.1858
#>             rs11265593 rs11265594 rs10908831 rs6686467 rs7555891 rs10218433
#> rs6691345       0.5624    -0.1369    -0.3042    0.8709    0.5624     0.5624
#> rs115165870    -0.1364    -0.1212    -0.1171   -0.0531   -0.1364    -0.1364
#> rs12727865      0.3394     0.6096    -0.5663   -0.1834    0.3394     0.3394
#> rs147889978    -0.1180    -0.1018    -0.1448   -0.0490   -0.1180    -0.1180
#> rs6702754       0.3394     0.6096    -0.5663   -0.1834    0.3394     0.3394
#> rs1209264       0.2411     0.1260     0.2965    0.1842    0.2411     0.2411
#>             rs35229198 rs11576181 rs12143987 rs11576193 rs11590941 rs4414034
#> rs6691345      -0.1388     0.5624    -0.5597     0.5638     0.5128   -0.1302
#> rs115165870    -0.1220    -0.1364     0.1357    -0.1355    -0.0646   -0.1220
#> rs12727865      0.6047     0.3394    -0.3378     0.3375    -0.1214    0.6047
#> rs147889978    -0.1027    -0.1180     0.1175    -0.1172    -0.0303   -0.1027
#> rs6702754       0.6047     0.3394    -0.3378     0.3375    -0.1214    0.6047
#> rs1209264       0.1187     0.2411    -0.2328     0.2395     0.0849    0.1277
#>             rs145030659 rs1073907 rs6698881 rs6696089 rs138273059 rs11577266
#> rs6691345        0.1349    0.5624    0.8709    0.5624     -0.0286     0.5624
#> rs115165870     -0.0114   -0.1364   -0.0531   -0.1364     -0.0037    -0.1364
#> rs12727865      -0.0591    0.3394   -0.1834    0.3394      0.1921     0.3394
#> rs147889978     -0.0098   -0.1180   -0.0490   -0.1180      0.7298    -0.1180
#> rs6702754       -0.0591    0.3394   -0.1834    0.3394      0.1921     0.3394
#> rs1209264        0.0765    0.2411    0.1842    0.2411      0.0665     0.2411
#>             rs11590203 rs144063783 rs6687939 rs6687971 rs1948143 rs72633646
#> rs6691345       0.8709     -0.0273   -0.1878   -0.1878   -0.1379     0.8709
#> rs115165870    -0.0531      0.0126    0.3114    0.3114   -0.1179    -0.0531
#> rs12727865     -0.1834     -0.0739    0.0974    0.0974    0.6045    -0.1834
#> rs147889978    -0.0490     -0.0401    0.3226    0.3226   -0.0984    -0.0490
#> rs6702754      -0.1834     -0.0739    0.0974    0.0974    0.6045    -0.1834
#> rs1209264       0.1842      0.0768   -0.6036   -0.6036    0.1282     0.1842
#>             rs12118634 rs150550224 rs12738097 rs12719998 rs72633647 rs6657938
#> rs6691345      -0.5585      0.1349     0.5599     0.5641     0.8709    0.5641
#> rs115165870     0.1347     -0.0114    -0.1212    -0.1336    -0.0531   -0.1336
#> rs12727865     -0.3308     -0.0591     0.3289     0.3335    -0.1834    0.3335
#> rs147889978     0.1164     -0.0098    -0.1155    -0.1152    -0.0490   -0.1152
#> rs6702754      -0.3308     -0.0591     0.3289     0.3335    -0.1834    0.3335
#> rs1209264      -0.2379      0.0765     0.2434     0.2429     0.1842    0.2429
#>             rs112559935 rs67156297 rs6698130 rs6698143 rs6664039 rs12749618
#> rs6691345       -0.0828    -0.1878    0.5641    0.5641    0.8709    -0.1379
#> rs115165870     -0.0411     0.3114   -0.1336   -0.1336   -0.0531    -0.1179
#> rs12727865      -0.1333     0.0974    0.3335    0.3335   -0.1834     0.6045
#> rs147889978     -0.0086     0.3226   -0.1152   -0.1152   -0.0490    -0.0984
#> rs6702754       -0.1333     0.0974    0.3335    0.3335   -0.1834     0.6045
#> rs1209264        0.0662    -0.6036    0.2429    0.2429    0.1842     0.1282
#>             rs12068128 rs6662976 rs113624284 rs116568035 rs68031662 rs35982711
#> rs6691345      -0.1379   -0.1379      0.8709     -0.0950    -0.1379     0.5641
#> rs115165870    -0.1179   -0.1179     -0.0531     -0.0837    -0.1179    -0.1336
#> rs12727865      0.6045    0.6045     -0.1834      0.3495     0.6045     0.3335
#> rs147889978    -0.0984   -0.0984     -0.0490     -0.0823    -0.0984    -0.1152
#> rs6702754       0.6045    0.6045     -0.1834      0.3495     0.6045     0.3335
#> rs1209264       0.1282    0.1282      0.1842      0.0479     0.1282     0.2429
#>             rs12063437 rs56807273 rs7552370 rs139364224 rs7515004 rs12025518
#> rs6691345      -0.1379     0.8709   -0.1379     -0.0658    0.5615    -0.1871
#> rs115165870    -0.1179    -0.0531   -0.1179     -0.0130   -0.1334     0.3138
#> rs12727865      0.6045    -0.1834    0.6045     -0.0497    0.3332     0.0961
#> rs147889978    -0.0984    -0.0490   -0.0984     -0.0510   -0.1150     0.3251
#> rs6702754       0.6045    -0.1834    0.6045     -0.0497    0.3332     0.0961
#> rs1209264       0.1282     0.1842    0.1282     -0.4059    0.2425    -0.6008
#>             rs12058332 rs6659828 rs7518394 rs79438587 rs6661909 rs6666888
#> rs6691345      -0.1379    0.5641   -0.1379    -0.1771    0.5494   -0.1379
#> rs115165870    -0.1179   -0.1336   -0.1179    -0.0842   -0.1325   -0.1179
#> rs12727865      0.6045    0.3335    0.6045    -0.3207    0.3362    0.6045
#> rs147889978    -0.0984   -0.1152   -0.0984    -0.0965   -0.1141   -0.0984
#> rs6702754       0.6045    0.3335    0.6045    -0.3207    0.3362    0.6045
#> rs1209264       0.1282    0.2429    0.1282     0.1290    0.2337    0.1282
#>             rs6662003 rs6701309 rs6667225 rs12061599 rs12066892 rs7511806
#> rs6691345     -0.1398   -0.1446   -0.1446     0.5588    -0.1465   -0.1465
#> rs115165870   -0.1187   -0.1170   -0.1170    -0.1327    -0.1179   -0.1179
#> rs12727865     0.6058    0.6033    0.6033     0.3316     0.5984    0.5984
#> rs147889978   -0.0829   -0.0976   -0.0976    -0.1144    -0.0984   -0.0984
#> rs6702754      0.6058    0.6033    0.6033     0.3316     0.5984    0.5984
#> rs1209264      0.1208    0.1265    0.1265     0.2342     0.1191    0.1191
#>             rs7525436 rs7523399 rs35221765 rs7530735 rs34055426 rs7530868
#> rs6691345      0.5573   -0.1465    -0.1976   -0.1465    -0.1976   -0.1465
#> rs115165870   -0.1336   -0.1179     0.3155   -0.1179     0.3155   -0.1179
#> rs12727865     0.3286    0.5984     0.1024    0.5984     0.1024    0.5984
#> rs147889978   -0.1152   -0.0984     0.3267   -0.0984     0.3267   -0.0984
#> rs6702754      0.3286    0.5984     0.1024    0.5984     0.1024    0.5984
#> rs1209264      0.2286    0.1191    -0.5893    0.1191    -0.5893    0.1191
#>             rs72696290 rs9727988 rs138765671 rs9728585 rs12029292 rs9728024
#> rs6691345      -0.3283   -0.1499      0.8639    0.5573    -0.3283   -0.1465
#> rs115165870    -0.1560   -0.1069     -0.0526   -0.1336    -0.1560   -0.1179
#> rs12727865     -0.3886    0.5785     -0.1819    0.3286    -0.3886    0.5984
#> rs147889978    -0.1825   -0.1036     -0.0487   -0.1152    -0.1825   -0.0984
#> rs6702754      -0.3886    0.5785     -0.1819    0.3286    -0.3886    0.5984
#> rs1209264       0.3167    0.1302      0.1735    0.2286     0.3167    0.1191
#>             rs4406622 rs6662855 rs6663105 rs6672627 rs9803896 rs4526605
#> rs6691345     -0.1465   -0.1465   -0.1465   -0.1465    0.8339   -0.1465
#> rs115165870   -0.1179   -0.1179   -0.1179   -0.1179   -0.0588   -0.1179
#> rs12727865     0.5984    0.5984    0.5984    0.5984   -0.2061    0.5984
#> rs147889978   -0.0984   -0.0984   -0.0984   -0.0984   -0.0548   -0.0984
#> rs6702754      0.5984    0.5984    0.5984    0.5984   -0.2061    0.5984
#> rs1209264      0.1191    0.1191    0.1191    0.1191    0.1292    0.1191
#>             rs4622059 rs58548028 rs61803412 rs28530583 rs28410194 rs9804073
#> rs6691345     -0.1427    -0.1708    -0.1446    -0.1446    -0.1446    0.8401
#> rs115165870   -0.1162    -0.0254    -0.1170    -0.1170    -0.1170   -0.0593
#> rs12727865     0.5897    -0.1710     0.5971     0.5971     0.5971   -0.1956
#> rs147889978   -0.0967    -0.0615    -0.0976    -0.0976    -0.0976   -0.0552
#> rs6702754      0.5897    -0.1710     0.5971     0.5971     0.5971   -0.1956
#> rs1209264      0.1156    -0.9243     0.1173     0.1173     0.1173    0.1302
#>             rs56264602 rs59356432 rs10908834 rs12565279 rs11582433 rs6690468
#> rs6691345       0.8435     0.8435    -0.5528    -0.1050     0.8435   -0.1446
#> rs115165870    -0.0583    -0.0583     0.1377     0.4225    -0.0583   -0.1170
#> rs12727865     -0.1980    -0.1980    -0.3083     0.2887    -0.1980    0.5971
#> rs147889978    -0.0543    -0.0543     0.1194     0.4670    -0.0543   -0.0976
#> rs6702754      -0.1980    -0.1980    -0.3083     0.2887    -0.1980    0.5971
#> rs1209264       0.1285     0.1285    -0.1870     0.0839     0.1285    0.1173
#>             rs61803414 rs3811451 rs16865579 rs34437053 rs57100877 rs60767732
#> rs6691345      -0.1446   -0.1446    -0.1446     0.5568     0.8435     0.8435
#> rs115165870    -0.1170   -0.1170    -0.1170    -0.1375    -0.0583    -0.0583
#> rs12727865      0.5971    0.5971     0.5971     0.3079    -0.1980    -0.1980
#> rs147889978    -0.0976   -0.0976    -0.0976    -0.1191    -0.0543    -0.0543
#> rs6702754       0.5971    0.5971     0.5971     0.3079    -0.1980    -0.1980
#> rs1209264       0.1173    0.1173     0.1173     0.1934     0.1285     0.1285
#>             rs60682501 rs60412881 rs1889313 rs1889314 rs61803415 rs1889315
#> rs6691345       0.8435     0.8435    0.8401   -0.5528    -0.1050    0.5568
#> rs115165870    -0.0583    -0.0583   -0.0593    0.1377     0.4225   -0.1375
#> rs12727865     -0.1980    -0.1980   -0.2016   -0.3083     0.2887    0.3079
#> rs147889978    -0.0543    -0.0543   -0.0552    0.1194     0.4670   -0.1191
#> rs6702754      -0.1980    -0.1980   -0.2016   -0.3083     0.2887    0.3079
#> rs1209264       0.1285     0.1285    0.1302   -0.1870     0.0839    0.1934
#>             rs61803416 rs138227751 rs1889316 rs4304577 rs35112510 rs36029470
#> rs6691345      -0.1050     -0.0572   -0.1446   -0.1446    -0.1446    -0.1379
#> rs115165870     0.4225      0.5873   -0.1170   -0.1170    -0.1170    -0.1179
#> rs12727865      0.2887      0.0700    0.5971    0.5971     0.5971     0.5922
#> rs147889978     0.4670      0.6428   -0.0976   -0.0976    -0.0976    -0.0984
#> rs6702754       0.2887      0.0700    0.5971    0.5971     0.5971     0.5922
#> rs1209264       0.0839      0.0669    0.1173    0.1173     0.1173     0.1191
#>             rs35629421 rs35091500 rs11265604 rs34749203 rs11580178 rs11579718
#> rs6691345      -0.1446    -0.1446    -0.3275    -0.1427     0.5568     0.8435
#> rs115165870    -0.1170    -0.1170    -0.1556    -0.1162    -0.1375    -0.0583
#> rs12727865      0.5971     0.5971    -0.3867     0.5897     0.3079    -0.1980
#> rs147889978    -0.0976    -0.0976    -0.1822    -0.0967    -0.1191    -0.0543
#> rs6702754       0.5971     0.5971    -0.3867     0.5897     0.3079    -0.1980
#> rs1209264       0.1173     0.1173     0.3161     0.1156     0.1934     0.1285
#>             rs35595831 rs34065906 rs112527081 rs116088025 rs61804488 rs61804489
#> rs6691345      -0.1446    -0.1446     -0.1446      0.8468    -0.1446    -0.1446
#> rs115165870    -0.1170    -0.1170     -0.1170     -0.0574    -0.1170    -0.1170
#> rs12727865      0.5971     0.5971      0.5971     -0.2006     0.5971     0.5971
#> rs147889978    -0.0976    -0.0976     -0.0976     -0.0534    -0.0976    -0.0976
#> rs6702754       0.5971     0.5971      0.5971     -0.2006     0.5971     0.5971
#> rs1209264       0.1173     0.1173      0.1173      0.1268     0.1173     0.1173
#>             rs56211693 rs12121085 rs78594767 rs72696301 rs186303011 rs12127600
#> rs6691345       0.5568    -0.3283    -0.0207    -0.1793     -0.1252    -0.3936
#> rs115165870    -0.1375    -0.1560    -0.0114    -0.0242     -0.0017     0.1436
#> rs12727865      0.3031    -0.3842    -0.0591    -0.1668     -0.1347    -0.1954
#> rs147889978    -0.1191    -0.1825    -0.0098    -0.0605     -0.0474     0.1485
#> rs6702754       0.3031    -0.3842    -0.0591    -0.1668     -0.1347    -0.1954
#> rs1209264       0.1934     0.3167     0.0765    -0.9299     -0.7230     0.4045
#>             rs11265606 rs10908835 rs61812563 rs11265607 rs4845612 rs6427627
#> rs6691345       0.5326     0.3976    -0.1031     0.5355   -0.3034    0.3976
#> rs115165870    -0.1107    -0.1424     0.4242    -0.1345   -0.1587   -0.1424
#> rs12727865      0.3101     0.1970     0.2932     0.3160   -0.3905    0.1970
#> rs147889978    -0.1178    -0.1473     0.4688    -0.1161   -0.1851   -0.1473
#> rs6702754       0.3101     0.1970     0.2932     0.3160   -0.3905    0.1970
#> rs1209264       0.1904    -0.3945     0.0914     0.1872    0.3219   -0.3945
#>             rs9651053 rs114800510 rs9651036 rs9651054 rs9651055 rs72698103
#> rs6691345     -0.1793     -0.0587   -0.1793    0.3976   -0.1793    -0.3034
#> rs115165870   -0.0242      0.0054   -0.0242   -0.1424   -0.0242    -0.1587
#> rs12727865    -0.1668      0.1405   -0.1668    0.1970   -0.1668    -0.3905
#> rs147889978   -0.0605     -0.0279   -0.0605   -0.1473   -0.0605    -0.1851
#> rs6702754     -0.1668      0.1405   -0.1668    0.1970   -0.1668    -0.3905
#> rs1209264     -0.9299      0.0624   -0.9299   -0.3945   -0.9299     0.3219
#>             rs72633650 rs115437600 rs111920902 rs12735458 rs12026876 rs77609234
#> rs6691345       0.8173     -0.0964     -0.1793    -0.0603    -0.1793    -0.1381
#> rs115165870    -0.0560     -0.0181     -0.0242    -0.0425    -0.0242    -0.0626
#> rs12727865     -0.1765     -0.1601     -0.1668     0.1316    -0.1668     0.4175
#> rs147889978    -0.0520      0.0089     -0.0605    -0.0417    -0.0605    -0.0825
#> rs6702754      -0.1765     -0.1601     -0.1668     0.1316    -0.1668     0.4175
#> rs1209264       0.1243      0.1021     -0.9299    -0.0647    -0.9299     0.0643
#>             rs12563459 rs151220053 rs61812565 rs68138282 rs11582424 rs11265608
#> rs6691345      -0.1031     -0.1793    -0.1419    -0.1419     0.5355    -0.1793
#> rs115165870     0.4242     -0.0242    -0.1163    -0.1163    -0.1345    -0.0242
#> rs12727865      0.2932     -0.1668     0.5869     0.5869     0.3160    -0.1668
#> rs147889978     0.4688     -0.0605    -0.0967    -0.0967    -0.1161    -0.0605
#> rs6702754       0.2932     -0.1668     0.5869     0.5869     0.3160    -0.1668
#> rs1209264       0.0914     -0.9299     0.1147     0.1147     0.1872    -0.9299
#>             rs11580535 rs11265609 rs9662562 rs12029939 rs9724691 rs12033701
#> rs6691345       0.8173     0.5355    0.5355    -0.3034    0.5355    -0.1793
#> rs115165870    -0.0560    -0.1345   -0.1345    -0.1587   -0.1345    -0.0242
#> rs12727865     -0.1765     0.3160    0.3160    -0.3905    0.3160    -0.1668
#> rs147889978    -0.0520    -0.1161   -0.1161    -0.1851   -0.1161    -0.0605
#> rs6702754      -0.1765     0.3160    0.3160    -0.3905    0.3160    -0.1668
#> rs1209264       0.1243     0.1872    0.1872     0.3219    0.1872    -0.9299
#>             rs6692515 rs34094138 rs7551007 rs113057497 rs11585004 rs138582043
#> rs6691345     -0.3010    -0.1031   -0.1031     -0.1793     0.3976     -0.1793
#> rs115165870   -0.1546     0.4242    0.4242     -0.0242    -0.1424     -0.0242
#> rs12727865    -0.3920     0.2932    0.2932     -0.1668     0.1970     -0.1668
#> rs147889978   -0.1811     0.4688    0.4688     -0.0605    -0.1473     -0.0605
#> rs6702754     -0.3920     0.2932    0.2932     -0.1668     0.1970     -0.1668
#> rs1209264      0.3142     0.0914    0.0914     -0.9299    -0.3945     -0.9299
#>             rs147763778 rs144111494 rs952146 rs17654071 rs111810442 rs73026617
#> rs6691345        0.4745     -0.1793   0.3976    -0.3034     -0.1793    -0.1793
#> rs115165870     -0.0349     -0.0242  -0.1424    -0.1587     -0.0242    -0.0242
#> rs12727865      -0.0555     -0.1668   0.1970    -0.3905     -0.1668    -0.1668
#> rs147889978     -0.0327     -0.0605  -0.1473    -0.1851     -0.0605    -0.0605
#> rs6702754       -0.0555     -0.1668   0.1970    -0.3905     -0.1668    -0.1668
#> rs1209264        0.1045     -0.9299  -0.3945     0.3219     -0.9299    -0.9299
#>             rs6427631 rs2054855 rs76033854 rs12075836 rs115717606 rs7553271
#> rs6691345      0.5341   -0.0766    -0.0712    -0.0646     -0.0611   -0.3477
#> rs115165870   -0.1226    0.4064    -0.0547     0.4219     -0.0089    0.1321
#> rs12727865     0.3179    0.2256     0.3077     0.2414      0.1614   -0.2214
#> rs147889978   -0.1040    0.2762    -0.0530     0.4679     -0.0508    0.1484
#> rs6702754      0.3179    0.2256     0.3077     0.2414      0.1614   -0.2214
#> rs1209264      0.1888    0.0658     0.0656     0.1216      0.0733    0.4102
#>             rs79207650 rs116179745 rs115200978 rs3887104 rs66654715 rs12132326
#> rs6691345       0.0205     -0.0877      0.3478   -0.0254    -0.0593    -0.0990
#> rs115165870    -0.0606     -0.0425     -0.0472    0.3700    -0.0886    -0.0411
#> rs12727865     -0.1588      0.1512     -0.0385    0.1750     0.3403    -0.1102
#> rs147889978    -0.0313     -0.0417      0.0007    0.4461    -0.0609    -0.0394
#> rs6702754      -0.1588      0.1512     -0.0385    0.1750     0.3403    -0.1102
#> rs1209264       0.0890      0.0222      0.0367    0.1510     0.0798     0.0662
#>             rs4845617 rs72698115 rs57569414 rs6427641 rs12083537 rs1386821
#> rs6691345      0.3513    -0.1812     0.7848   -0.3659    -0.0825   -0.0880
#> rs115165870   -0.1325    -0.0254    -0.0411    0.1514    -0.1336   -0.1303
#> rs12727865     0.2179    -0.1636    -0.1946   -0.1740     0.4881    0.5065
#> rs147889978   -0.1487    -0.0615    -0.0372    0.1323    -0.1010   -0.0972
#> rs6702754      0.2179    -0.1636    -0.1946   -0.1740     0.4881    0.5065
#> rs1209264     -0.4076    -0.9133     0.0874    0.3645     0.1312    0.1230
#>             rs12135008 rs4521987 rs4075015 rs11588248 rs79794939 rs35717427
#> rs6691345      -0.0550    0.0570   -0.3215    -0.0295    -0.0121     0.7508
#> rs115165870    -0.0362   -0.4123   -0.1288     0.3669     0.5544    -0.0430
#> rs12727865     -0.1754   -0.2053   -0.3211     0.1544     0.0395    -0.1829
#> rs147889978    -0.0345   -0.4405   -0.1560     0.4265     0.6123    -0.0223
#> rs6702754      -0.1754   -0.2053   -0.3211     0.1544     0.0395    -0.1829
#> rs1209264       0.0375   -0.1209    0.2814     0.1454     0.0838     0.0817
#>             rs146389272 rs4556347 rs4556348 rs7525477 rs4601580 rs12047973
#> rs6691345        0.2049   -0.2581   -0.0975   -0.2344    0.2033    -0.1286
#> rs115165870      0.0240   -0.1704   -0.0423    0.0270   -0.0330    -0.0388
#> rs12727865      -0.1097    0.0849   -0.0961    0.0756   -0.0942    -0.1304
#> rs147889978     -0.0050   -0.1936   -0.1243    0.1232   -0.1173    -0.0965
#> rs6702754       -0.1097    0.0849   -0.0961    0.0756   -0.0942    -0.1304
#> rs1209264       -0.0097   -0.1381   -0.2720    0.2725   -0.2549    -0.2743
#>             rs12048950 rs12048091 rs11265611 rs4133213 rs73026642 rs6684439
#> rs6691345      -0.1223    -0.1281     0.2154    0.3089    -0.0330    0.3094
#> rs115165870    -0.0398     0.0537    -0.0908   -0.0599     0.3169   -0.0629
#> rs12727865     -0.1340    -0.1199    -0.1333   -0.0413     0.0338   -0.0764
#> rs147889978    -0.0974    -0.0913    -0.1266   -0.0515     0.0104   -0.0430
#> rs6702754      -0.1340    -0.1199    -0.1333   -0.0413     0.0338   -0.0764
#> rs1209264      -0.2809    -0.2825    -0.0166    0.1845    -0.0357    0.1869
#>             rs6689306 rs77302258 rs12118721 rs12118770 rs12117832 rs10908838
#> rs6691345      0.1988    -0.0973     0.2075     0.1988     0.2075     0.1998
#> rs115165870   -0.0199    -0.0213    -0.0916    -0.0199    -0.0916    -0.0191
#> rs12727865    -0.1653     0.0032    -0.1744    -0.1653    -0.1744    -0.1665
#> rs147889978   -0.1143    -0.0857    -0.1157    -0.1143    -0.1157    -0.1133
#> rs6702754     -0.1653     0.0032    -0.1744    -0.1653    -0.1744    -0.1665
#> rs1209264     -0.0375    -0.4926    -0.0241    -0.0375    -0.0241    -0.0322
#>             rs111266048 rs78186112 rs12023358 rs4845618 rs6687726 rs73014267
#> rs6691345       -0.0330    -0.0330    -0.1267    0.2075    0.2075    -0.0330
#> rs115165870      0.3169     0.3169    -0.0379   -0.0916   -0.0916     0.3169
#> rs12727865       0.0338     0.0338    -0.1326   -0.1744   -0.1744     0.0338
#> rs147889978      0.0104     0.0104    -0.0956   -0.1157   -0.1157     0.0104
#> rs6702754        0.0338     0.0338    -0.1326   -0.1744   -0.1744     0.0338
#> rs1209264       -0.0357    -0.0357    -0.2765   -0.0241   -0.0241    -0.0357
#>             rs6427658 rs73014271 rs58281895 rs2228144 rs2229237 rs6694817
#> rs6691345      0.1998    -0.0330    -0.0330   -0.1267   -0.0330    0.1998
#> rs115165870   -0.0191     0.3169     0.3169   -0.0379    0.3169   -0.0191
#> rs12727865    -0.1665     0.0338     0.0338   -0.1326    0.0338   -0.1665
#> rs147889978   -0.1133     0.0104     0.0104   -0.0956    0.0104   -0.1133
#> rs6702754     -0.1665     0.0338     0.0338   -0.1326    0.0338   -0.1665
#> rs1209264     -0.0322    -0.0357    -0.0357   -0.2765   -0.0357   -0.0322
#>             rs73014279 rs7549250 rs7549338 rs7553796 rs116037345 rs56383622
#> rs6691345      -0.0330    0.2018    0.1942    0.1942     -0.0793     0.3005
#> rs115165870     0.3169   -0.0933   -0.0212   -0.0212      0.0336    -0.0535
#> rs12727865      0.0338   -0.1776   -0.1698   -0.1698      0.2392    -0.0871
#> rs147889978     0.0104   -0.1055   -0.1031   -0.1031     -0.0445    -0.0336
#> rs6702754       0.0338   -0.1776   -0.1698   -0.1698      0.2392    -0.0871
#> rs1209264      -0.0357   -0.0198   -0.0279   -0.0279      0.0621     0.1843
#>             rs4845619 rs115617158 rs59632925 rs4845620 rs7518199 rs7521458
#> rs6691345      0.2018     -0.0432     0.1942    0.3005    0.3005    0.2993
#> rs115165870   -0.0933     -0.0221    -0.0212   -0.0535   -0.0535   -0.0543
#> rs12727865    -0.1776      0.1584    -0.1698   -0.0871   -0.0871   -0.0856
#> rs147889978   -0.1055      0.6261    -0.1031   -0.0336   -0.0336   -0.0343
#> rs6702754     -0.1776      0.1584    -0.1698   -0.0871   -0.0871   -0.0856
#> rs1209264     -0.0198      0.0490    -0.0279    0.1843    0.1843    0.1859
#>             rs4845371 rs57502626 rs6667434 rs79778789 rs55826755 rs4845621
#> rs6691345      0.1942    -0.1267    0.1942     0.2521    -0.1248    0.3005
#> rs115165870   -0.0212    -0.0534   -0.0212    -0.0441    -0.0524   -0.0535
#> rs12727865    -0.1698    -0.1208   -0.1698     0.0018    -0.1290   -0.0871
#> rs147889978   -0.1031    -0.0956   -0.1031     0.0572    -0.0948   -0.0336
#> rs6702754     -0.1698    -0.1208   -0.1698     0.0018    -0.1290   -0.0871
#> rs1209264     -0.0279    -0.2678   -0.0279     0.0832    -0.2699    0.1843
#>             rs61812592 rs61812593 rs4553185 rs4845622 rs115224285 rs61812594
#> rs6691345      -0.1267    -0.1267    0.2018    0.3005     -0.0570    -0.1185
#> rs115165870    -0.0534    -0.0534   -0.0933   -0.0535     -0.0502    -0.0534
#> rs12727865     -0.1208    -0.1208   -0.1776   -0.0871     -0.0577    -0.1326
#> rs147889978    -0.0956    -0.0956   -0.1055   -0.0336      0.0396    -0.0956
#> rs6702754      -0.1208    -0.1208   -0.1776   -0.0871     -0.0577    -0.1326
#> rs1209264      -0.2678    -0.2678   -0.0198    0.1843      0.0452    -0.2765
#>             rs76093405 rs61812595 rs59502179 rs149219382 rs113325045
#> rs6691345      -0.0432    -0.0576    -0.0330     -0.0429     -0.0811
#> rs115165870    -0.0221    -0.0211     0.3169     -0.0114     -0.0319
#> rs12727865      0.1584    -0.1605     0.0338      0.1632     -0.0839
#> rs147889978     0.6261    -0.0371     0.0104      0.7094      0.0030
#> rs6702754       0.1584    -0.1605     0.0338      0.1632     -0.0839
#> rs1209264       0.0490     0.0906    -0.0357      0.0530      0.0479
#>             rs114660934 rs4393147 rs4453032 rs6664201 rs61812596 rs4845372
#> rs6691345       -0.1098    0.3005    0.3005    0.3005    -0.1248    0.2908
#> rs115165870     -0.0381   -0.0535   -0.0535   -0.0535    -0.0524    0.0191
#> rs12727865      -0.0860   -0.0871   -0.0871   -0.0871    -0.1290   -0.0699
#> rs147889978     -0.0365   -0.0336   -0.0336   -0.0336    -0.0948   -0.0310
#> rs6702754       -0.0860   -0.0871   -0.0871   -0.0871    -0.1290   -0.0699
#> rs1209264       -0.3952    0.1843    0.1843    0.1843    -0.2699    0.1748
#>             rs4845623 rs116141616 rs139952834 rs12753254 rs12730036 rs11265612
#> rs6691345      0.2908      0.0833     -0.0926     0.3030     0.3030     0.1971
#> rs115165870    0.0191     -0.0064     -0.0285    -0.0519    -0.0519    -0.0197
#> rs12727865    -0.0699      0.0259     -0.1471    -0.0857    -0.0857    -0.1684
#> rs147889978   -0.0310     -0.0049     -0.0270    -0.0320    -0.0320    -0.1015
#> rs6702754     -0.0699      0.0259     -0.1471    -0.0857    -0.0857    -0.1684
#> rs1209264      0.1748      0.0452      0.0032     0.1811     0.1811    -0.0307
#>             rs55800510 rs4845373 rs6683206 rs113617296 rs186043699 rs11265613
#> rs6691345      -0.1248    0.3018    0.1971     -0.0330     -0.0763     0.2813
#> rs115165870    -0.0524   -0.0527   -0.0197      0.3169     -0.0571    -0.0607
#> rs12727865     -0.1290   -0.0842   -0.1640      0.0338      0.0055    -0.0841
#> rs147889978    -0.0948   -0.0328   -0.1015      0.0104     -0.0165    -0.0528
#> rs6702754      -0.1290   -0.0842   -0.1640      0.0338      0.0055    -0.0841
#> rs1209264      -0.2699    0.1761   -0.0241     -0.0357      0.0416     0.1632
#>             rs4576655 rs4537545 rs11265614 rs58229059 rs111301013 rs6686750
#> rs6691345      0.3031    0.2981    -0.0273    -0.0383     -0.1284    0.2033
#> rs115165870    0.0238    0.0246     0.3329     0.3025     -0.0502   -0.0138
#> rs12727865    -0.0759   -0.0730     0.0074     0.0018     -0.1209   -0.1726
#> rs147889978   -0.0268   -0.0261     0.0141     0.0069     -0.0924   -0.0957
#> rs6702754     -0.0759   -0.0730     0.0074     0.0018     -0.1209   -0.1726
#> rs1209264      0.1748    0.1732    -0.0436    -0.0285     -0.2549   -0.0221
#>             rs61812598 rs113580743 rs7512646 rs57342848 rs7529229 rs111808975
#> rs6691345       0.3145     -0.0698    0.3044    -0.0273    0.3044     -0.0273
#> rs115165870    -0.0488     -0.0425    0.0246     0.3329    0.0246      0.3329
#> rs12727865     -0.0905     -0.0705   -0.0820     0.0074   -0.0820      0.0074
#> rs147889978    -0.0287      0.0200   -0.0261     0.0141   -0.0261      0.0141
#> rs6702754      -0.0905     -0.0705   -0.0820     0.0074   -0.0820      0.0074
#> rs1209264       0.1829      0.0690    0.1732    -0.0436    0.1732     -0.0436
#>             rs4845625 rs45478197 rs4845626 rs12129500 rs7536152 rs11804305
#> rs6691345      0.2043    -0.1019   -0.1301     0.2043    0.2043    -0.1265
#> rs115165870   -0.0130    -0.0399    0.0454    -0.0130   -0.0130    -0.0493
#> rs12727865    -0.1693     0.0196   -0.1115    -0.1693   -0.1693    -0.1173
#> rs147889978   -0.0948    -0.0828   -0.0846    -0.0948   -0.0948    -0.0915
#> rs6702754     -0.1693     0.0196   -0.1115    -0.1693   -0.1693    -0.1173
#> rs1209264     -0.0169    -0.4577   -0.2526    -0.0169   -0.0169    -0.2483
#>             rs7526131 rs12126142 rs6689393 rs12739228 rs4129267 rs4845374
#> rs6691345      0.2043     0.3095    0.2000    -0.0581    0.3095   -0.1284
#> rs115165870   -0.0130    -0.0480   -0.0123    -0.0869   -0.0480   -0.0502
#> rs12727865    -0.1693    -0.0921   -0.1712     0.2952   -0.0921   -0.1092
#> rs147889978   -0.0948    -0.0279   -0.0942    -0.0597   -0.0279   -0.0924
#> rs6702754     -0.1693    -0.0921   -0.1712     0.2952   -0.0921   -0.1092
#> rs1209264     -0.0169     0.1813   -0.0183     0.0640    0.1813   -0.2462
#>             rs2228145 rs12133641 rs140393337 rs6694258 rs11265618 rs191793127
#> rs6691345      0.3095     0.3056     -0.0252    0.2000    -0.1282     -0.0790
#> rs115165870   -0.0480    -0.0494     -0.0377   -0.0123     0.0464     -0.0757
#> rs12727865    -0.0921    -0.0840     -0.0307   -0.1712    -0.1137      0.2754
#> rs147889978   -0.0279    -0.0293     -0.0370   -0.0942    -0.0838     -0.0450
#> rs6702754     -0.0921    -0.0840     -0.0307   -0.1712    -0.1137      0.2754
#> rs1209264      0.1813     0.1837      0.0710   -0.0183    -0.2546      0.0445
#>             rs10908839 rs142712385 rs151017214 rs10159236 rs10752641 rs4329505
#> rs6691345       0.1849     -0.0762     -0.0762    -0.1227     0.1918   -0.1246
#> rs115165870     0.1032     -0.0202     -0.0202    -0.0474     0.0239   -0.0484
#> rs12727865      0.0366     -0.0204     -0.0204    -0.1159     0.0377   -0.1078
#> rs147889978     0.0197     -0.0926     -0.0926    -0.0898     0.0170   -0.0906
#> rs6702754       0.0366     -0.0204     -0.0204    -0.1159     0.0377   -0.1078
#> rs1209264       0.0866     -0.3928     -0.3928    -0.2525     0.1021   -0.2504
#>             rs4308966 rs6690230 rs6698040 rs6695045 rs7537291 rs7546068
#> rs6691345     -0.1246    0.2650    0.1918    0.2650    0.1918    0.1918
#> rs115165870   -0.0484    0.0574    0.0239    0.0574    0.0239    0.0239
#> rs12727865    -0.1078    0.1138    0.0377    0.1138    0.0377    0.0377
#> rs147889978   -0.0906    0.0825    0.0170    0.0825    0.0170    0.0170
#> rs6702754     -0.1078    0.1138    0.0377    0.1138    0.0377    0.0377
#> rs1209264     -0.2504    0.2776    0.1021    0.2776    0.1021    0.1021
#>             rs7537316 rs7546552 rs7546555 rs35938325 rs114879247 rs6687597
#> rs6691345      0.1918    0.1918    0.1918    -0.0330     -0.0384    0.1918
#> rs115165870    0.0239    0.0239    0.0239     0.3169      0.0338    0.0239
#> rs12727865     0.0377    0.0377    0.0377    -0.0053      0.1125    0.0377
#> rs147889978    0.0170    0.0170    0.0170     0.0104     -0.0074    0.0170
#> rs6702754      0.0377    0.0377    0.0377    -0.0053      0.1125    0.0377
#> rs1209264      0.1021    0.1021    0.1021    -0.0647      0.0251    0.1021
#>             rs55676222 rs55668699 rs6427672 rs4240872 rs4509570 rs4341355
#> rs6691345      -0.0473    -0.1870    0.1918    0.1926    0.1926    0.1926
#> rs115165870    -0.0505    -0.1078    0.0239    0.0246    0.0246    0.0246
#> rs12727865     -0.1318    -0.0369    0.0377    0.0357    0.0357    0.0357
#> rs147889978     0.0129    -0.0093    0.0170    0.0177    0.0177    0.0177
#> rs6702754      -0.1318    -0.0369    0.0377    0.0357    0.0357    0.0357
#> rs1209264       0.1181    -0.0542    0.1021    0.1001    0.1001    0.1001
#>             rs61812626 rs4345796 rs4291493 rs2229238 rs7514452 rs114445392
#> rs6691345      -0.1227   -0.1227    0.1944    0.1568    0.1513      0.2407
#> rs115165870    -0.0474   -0.0474    0.0254    0.1241    0.1249     -0.0425
#> rs12727865     -0.1159   -0.1159    0.0340    0.1544    0.1525     -0.0053
#> rs147889978    -0.0898   -0.0898    0.0185   -0.0053   -0.0044      0.0625
#> rs6702754      -0.1159   -0.1159    0.0340    0.1544    0.1525     -0.0053
#> rs1209264      -0.2525   -0.2525    0.0986    0.1158    0.1141      0.0800
#>             rs4072391 rs41310931 rs4379670 rs73016210 rs41308415 rs11265621
#> rs6691345      0.1607    -0.0481    0.1563    -0.0330    -0.0330     0.2384
#> rs115165870    0.1224     0.2777    0.1206     0.3169     0.3169     0.0681
#> rs12727865     0.1479    -0.0208    0.1599    -0.0053    -0.0053     0.2011
#> rs147889978    0.0069     0.0007    0.0184     0.0104     0.0104     0.0715
#> rs6702754      0.1479    -0.0208    0.1599    -0.0053    -0.0053     0.2011
#> rs1209264      0.1037    -0.0418    0.1171    -0.0647    -0.0647     0.3055
#>             rs7526293 rs6669229 rs79480105 rs6675472 rs58348886 rs150056239
#> rs6691345      0.1661    0.1661    -0.0985    0.1661    -0.1636     -0.1014
#> rs115165870    0.0366    0.0366    -0.0319    0.0366    -0.0351     -0.0658
#> rs12727865     0.1499    0.1499    -0.1214    0.1499    -0.1493      0.0571
#> rs147889978    0.0170    0.0170    -0.0635    0.0170    -0.0154     -0.0319
#> rs6702754      0.1499    0.1499    -0.1214    0.1499    -0.1493      0.0571
#> rs1209264      0.1291    0.1291    -0.3770    0.1291    -0.1255      0.0147
#>             rs11265622 rs14021 rs76518735 rs144029367 rs12568083 rs4474240
#> rs6691345       0.2366 -0.1218    -0.1014     -0.0985     0.2366    0.1275
#> rs115165870     0.0706 -0.0471    -0.0658     -0.0646     0.0706    0.0499
#> rs12727865      0.1976 -0.0976     0.0571      0.0659     0.1976    0.0909
#> rs147889978     0.0739 -0.0735    -0.0319     -0.0303     0.0739    0.0762
#> rs6702754       0.1976 -0.0976     0.0571      0.0659     0.1976    0.0909
#> rs1209264       0.3109 -0.2594     0.0147      0.0110     0.3109    0.2617
#>             rs10908840 rs111885536 rs41269913 rs16836054 rs12044132 rs115306743
#> rs6691345      -0.0663     -0.1218     0.1729    -0.1218     0.4082     -0.0453
#> rs115165870     0.0893     -0.0471     0.0037    -0.0471    -0.0263     -0.0445
#> rs12727865      0.2385     -0.0976    -0.0688    -0.0976    -0.0583      0.0137
#> rs147889978     0.1011     -0.0735    -0.0214    -0.0735    -0.0379      0.0449
#> rs6702754       0.2385     -0.0976    -0.0688    -0.0976    -0.0583      0.0137
#> rs1209264       0.2205     -0.2594     0.0242    -0.2594     0.1186      0.0592
#>             rs4845634 rs4478801 rs6684921 rs73018293 rs6686276 rs59838898
#> rs6691345     -0.1218    0.2366    0.2366    -0.1607    0.2366    -0.1607
#> rs115165870   -0.0471    0.0706    0.0706    -0.0375    0.0706    -0.0375
#> rs12727865    -0.0976    0.1976    0.1976    -0.1479    0.1976    -0.1479
#> rs147889978   -0.0735    0.0739    0.0739    -0.0179    0.0739    -0.0179
#> rs6702754     -0.0976    0.1976    0.1976    -0.1479    0.1976    -0.1479
#> rs1209264     -0.2594    0.3109    0.3109    -0.1275    0.3109    -0.1275
#>             rs61812631 rs10047079 rs6689710 rs6689965 rs56233546 rs6700296
#> rs6691345      -0.1218    -0.1607    0.1275    0.2366    -0.1218    0.2366
#> rs115165870    -0.0471    -0.0375    0.0499    0.0706    -0.0471    0.0706
#> rs12727865     -0.0976    -0.1479    0.0909    0.1976    -0.0976    0.1976
#> rs147889978    -0.0735    -0.0179    0.0762    0.0739    -0.0735    0.0739
#> rs6702754      -0.0976    -0.1479    0.0909    0.1976    -0.0976    0.1976
#> rs1209264      -0.2594    -0.1275    0.2617    0.3109    -0.2594    0.3109
#>             rs4845635 rs6703672 rs12753666 rs12753680 rs6658175 rs6698971
#> rs6691345      0.1275    0.1275     0.2366     0.2366    0.2376    0.2376
#> rs115165870    0.0499    0.0499     0.0706     0.0706    0.0709    0.0709
#> rs12727865     0.0909    0.0909     0.1976     0.1976    0.1984    0.1984
#> rs147889978    0.0762    0.0762     0.0739     0.0739    0.0742    0.0742
#> rs6702754      0.0909    0.0909     0.1976     0.1976    0.1984    0.1984
#> rs1209264      0.2617    0.2617     0.3109     0.3109    0.3122    0.3122
#>             rs6662503 rs6696177 rs12118018 rs12118074 rs75456865 rs12124333
#> rs6691345      0.1275    0.1275     0.2366     0.1275    -0.0302     0.1275
#> rs115165870    0.0499    0.0499     0.0706     0.0499     0.0239     0.0499
#> rs12727865     0.0909    0.0909     0.1976     0.0909     0.1063     0.0909
#> rs147889978    0.0762    0.0762     0.0739     0.0762    -0.0144     0.0762
#> rs6702754      0.0909    0.0909     0.1976     0.0909     0.1063     0.0909
#> rs1209264      0.2617    0.2617     0.3109     0.2617     0.0153     0.2617
#>             rs12119111 rs146563764 rs6664608 rs35013837 rs72698167 rs66980031
#> rs6691345       0.2348     -0.0570    0.2366    -0.0702    -0.1218    -0.0660
#> rs115165870     0.0698     -0.0502    0.0706    -0.0244    -0.0471    -0.0593
#> rs12727865      0.1945     -0.0744    0.1976     0.2331    -0.0976     0.2731
#> rs147889978     0.0731     -0.0049    0.0739    -0.0733    -0.0735    -0.0575
#> rs6702754       0.1945     -0.0744    0.1976     0.2331    -0.0976     0.2731
#> rs1209264       0.3119      0.0205    0.3109     0.0500    -0.2594     0.0588
#>             rs7513603 rs10752642 rs73020232 rs73020234 rs116059394 rs12023772
#> rs6691345      0.1275     0.1275    -0.1607    -0.1607     -0.1014     0.4109
#> rs115165870    0.0499     0.0499    -0.0375    -0.0375     -0.0300    -0.0254
#> rs12727865     0.0909     0.0909    -0.1479    -0.1479     -0.1553    -0.0606
#> rs147889978    0.0762     0.0762    -0.0179    -0.0179     -0.0824    -0.0370
#> rs6702754      0.0909     0.0909    -0.1479    -0.1479     -0.1553    -0.0606
#> rs1209264      0.2617     0.2617    -0.1275    -0.1275     -0.3118     0.1169
#>             rs9660786 rs7518694 rs7526247 rs73020246 rs4845636 rs72698169
#> rs6691345      0.2366    0.2302    0.1275    -0.1607    0.1275     0.4109
#> rs115165870    0.0706    0.0676    0.0499    -0.0375    0.0499    -0.0254
#> rs12727865     0.1976    0.2083    0.0909    -0.1479    0.0909    -0.0606
#> rs147889978    0.0739    0.0947    0.0762    -0.0179    0.0762    -0.0370
#> rs6702754      0.1976    0.2083    0.0909    -0.1479    0.0909    -0.0606
#> rs1209264      0.3109    0.3096    0.2617    -0.1275    0.2617     0.1169
#>             rs12740969 rs11490956 rs111600849 rs12403537 rs10908841 rs149793513
#> rs6691345       0.2357    -0.1589     -0.1589    -0.1589     0.2357     -0.0273
#> rs115165870     0.0701    -0.0366     -0.0366    -0.0366     0.0701      0.3329
#> rs12727865      0.1998    -0.1499     -0.1499    -0.1499     0.1998      0.0074
#> rs147889978     0.0734    -0.0170     -0.0170    -0.0170     0.0734      0.0141
#> rs6702754       0.1998    -0.1499     -0.1499    -0.1499     0.1998      0.0074
#> rs1209264       0.3132    -0.1291     -0.1291    -0.1291     0.3132     -0.0436
#>             rs7519499 rs12128408 rs4845637 rs4845638 rs4845639 rs6674171
#> rs6691345      0.2357     0.2357    0.2357    0.2357    0.2357   -0.1589
#> rs115165870    0.0701     0.0701    0.0701    0.0701    0.0701   -0.0366
#> rs12727865     0.1998     0.1998    0.1998    0.1998    0.1998   -0.1499
#> rs147889978    0.0734     0.0734    0.0734    0.0734    0.0734   -0.0170
#> rs6702754      0.1998     0.1998    0.1998    0.1998    0.1998   -0.1499
#> rs1209264      0.3132     0.3132    0.3132    0.3132    0.3132   -0.1291
#>             rs59239860 rs4288587 rs4518898 rs77184252 rs10908842 rs10908843
#> rs6691345      -0.1516    0.1275    0.1275    -0.1589     0.2357     0.2339
#> rs115165870    -0.1216    0.0499    0.0499    -0.0366     0.0701     0.0692
#> rs12727865     -0.1499    0.0909    0.0909    -0.1499     0.1998     0.1967
#> rs147889978    -0.0200    0.0762    0.0762    -0.0170     0.0734     0.0726
#> rs6702754      -0.1499    0.0909    0.0909    -0.1499     0.1998     0.1967
#> rs1209264      -0.1131    0.2617    0.2617    -0.1291     0.3132     0.3142
#>             rs6670375 rs4575077 rs4345797 rs4584384 rs11265627 rs56100876
#> rs6691345      0.1275    0.1275    0.2339    0.2339     0.1275     0.0132
#> rs115165870    0.0499    0.0499    0.0692    0.0692     0.0499    -0.0064
#> rs12727865     0.0909    0.0909    0.1967    0.1967     0.0909    -0.0744
#> rs147889978    0.0762    0.0762    0.0726    0.0726     0.0762    -0.0494
#> rs6702754      0.0909    0.0909    0.1967    0.1967     0.0909    -0.0744
#> rs1209264      0.2617    0.2617    0.3142    0.3142     0.2617     0.0700
#>             rs11265628 rs6691727 rs6687112 rs6427716 rs4845640 rs4845641
#> rs6691345       0.2339    0.1275    0.1275    0.1275    0.1275    0.1275
#> rs115165870     0.0692    0.0499    0.0499    0.0499    0.0499    0.0499
#> rs12727865      0.1967    0.0909    0.0909    0.0909    0.0909    0.0909
#> rs147889978     0.0726    0.0762    0.0762    0.0762    0.0762    0.0762
#> rs6702754       0.1967    0.0909    0.0909    0.0909    0.0909    0.0909
#> rs1209264       0.3142    0.2617    0.2617    0.2617    0.2617    0.2617
#>             rs4845642 rs75631461 rs4845643 rs11265629 rs4585969 rs4307559
#> rs6691345      0.2339     0.0132    0.1275     0.1275    0.1275    0.1275
#> rs115165870    0.0692    -0.0064    0.0499     0.0499    0.0499    0.0499
#> rs12727865     0.1967    -0.0744    0.0909     0.0909    0.0909    0.0909
#> rs147889978    0.0726    -0.0494    0.0762     0.0762    0.0762    0.0762
#> rs6702754      0.1967    -0.0744    0.0909     0.0909    0.0909    0.0909
#> rs1209264      0.3142     0.0700    0.2617     0.2617    0.2617    0.2617
#>             rs78664422 rs4307560 rs12406822 rs6701586 rs6697103 rs6697115
#> rs6691345      -0.0278    0.1275    -0.1572    0.1275    0.1275    0.1275
#> rs115165870     0.1936    0.0499    -0.0358    0.0499    0.0499    0.0499
#> rs12727865      0.0999    0.0909    -0.1466    0.0909    0.0909    0.0909
#> rs147889978     0.2363    0.0762    -0.0162    0.0762    0.0762    0.0762
#> rs6702754       0.0999    0.0909    -0.1466    0.0909    0.0909    0.0909
#> rs1209264       0.0910    0.2617    -0.1308    0.2617    0.2617    0.2617
#>             rs6656395 rs4131514 rs10908844 rs10908845 rs12129484 rs10908846
#> rs6691345      0.1275    0.1275     0.1275     0.2339     0.1275     0.1275
#> rs115165870    0.0499    0.0499     0.0499     0.0692     0.0499     0.0499
#> rs12727865     0.0909    0.0909     0.0909     0.1967     0.0909     0.0909
#> rs147889978    0.0762    0.0762     0.0762     0.0726     0.0762     0.0762
#> rs6702754      0.0909    0.0909     0.0909     0.1967     0.0909     0.0909
#> rs1209264      0.2617    0.2617     0.2617     0.3142     0.2617     0.2617
#>             rs4845644 rs6427720 rs6427721 rs6427722 rs116710125 rs6427724
#> rs6691345     -0.1218    0.1275    0.1275    0.1275     -0.0066    0.1275
#> rs115165870   -0.0471    0.0499    0.0499    0.0499     -0.0053    0.0499
#> rs12727865    -0.0976    0.0909    0.0909    0.0909      0.0787    0.0909
#> rs147889978   -0.0735    0.0762    0.0762    0.0762     -0.0361    0.0762
#> rs6702754     -0.0976    0.0909    0.0909    0.0909      0.0787    0.0909
#> rs1209264     -0.2594    0.2617    0.2617    0.2617      0.0408    0.2617
#>             rs61812654 rs9427076 rs10752643 rs60498985 rs59741504 rs60255122
#> rs6691345      -0.1218   -0.1218     0.1275    -0.1572    -0.1572    -0.1572
#> rs115165870    -0.0471   -0.0471     0.0499    -0.0358    -0.0358    -0.0358
#> rs12727865     -0.0976   -0.0976     0.0909    -0.1466    -0.1466    -0.1466
#> rs147889978    -0.0735   -0.0735     0.0762    -0.0162    -0.0162    -0.0162
#> rs6702754      -0.0976   -0.0976     0.0909    -0.1466    -0.1466    -0.1466
#> rs1209264      -0.2594   -0.2594     0.2617    -0.1308    -0.1308    -0.1308
#>             rs61275241 rs112231452 rs61403567 rs10908847 rs58710275 rs77994623
#> rs6691345      -0.1572     -0.1572    -0.1572     0.2339    -0.1572    -0.0256
#> rs115165870    -0.0358     -0.0358    -0.0358     0.0692    -0.0358     0.1951
#> rs12727865     -0.1466     -0.1466    -0.1466     0.1967    -0.1466     0.1043
#> rs147889978    -0.0162     -0.0162    -0.0162     0.0726    -0.0162     0.2378
#> rs6702754      -0.1466     -0.1466    -0.1466     0.1967    -0.1466     0.1043
#> rs1209264      -0.1308     -0.1308    -0.1308     0.3142    -0.1308     0.0989
#>             rs12407048 rs4638123 rs7551873 rs7556449 rs41308419 rs112585956
#> rs6691345       0.2339    0.2330   -0.1572   -0.1572    -0.1572     -0.1572
#> rs115165870     0.0692    0.0687   -0.0358   -0.0358    -0.0358     -0.0358
#> rs12727865      0.1967    0.1944   -0.1466   -0.1466    -0.1466     -0.1466
#> rs147889978     0.0726    0.0720   -0.0162   -0.0162    -0.0162     -0.0162
#> rs6702754       0.1967    0.1944   -0.1466   -0.1466    -0.1466     -0.1466
#> rs1209264       0.3142    0.3165   -0.1308   -0.1308    -0.1308     -0.1308
#>             rs12037271 rs4633282 rs73023331 rs9427415 rs111742980 rs112674677
#> rs6691345       0.2339   -0.1572    -0.1572   -0.1218     -0.1572     -0.1572
#> rs115165870     0.0692   -0.0358    -0.0358   -0.0471     -0.0358     -0.0358
#> rs12727865      0.1967   -0.1466    -0.1466   -0.0976     -0.1466     -0.1466
#> rs147889978     0.0726   -0.0162    -0.0162   -0.0735     -0.0162     -0.0162
#> rs6702754       0.1967   -0.1466    -0.1466   -0.0976     -0.1466     -0.1466
#> rs1209264       0.3142   -0.1308    -0.1308   -0.2594     -0.1308     -0.1308
#>             rs113915076 rs4639752 rs147830103 rs4390168 rs60368585 rs116805289
#> rs6691345       -0.1572   -0.1572     -0.1099   -0.1517    -0.1572     -0.0570
#> rs115165870     -0.0358   -0.0358     -0.0757   -0.0366    -0.0358     -0.0064
#> rs12727865      -0.1466   -0.1466      0.1207   -0.1447    -0.1466      0.1595
#> rs147889978     -0.0162   -0.0162      0.1023   -0.0170    -0.0162      0.6627
#> rs6702754       -0.1466   -0.1466      0.1207   -0.1447    -0.1466      0.1595
#> rs1209264       -0.1308   -0.1308      0.0118   -0.1291    -0.1308      0.0452
#>             rs112236096 rs73023339 rs12136771 rs57783436 rs112394421 rs73023346
#> rs6691345       -0.1572    -0.1572     0.2293    -0.1519     -0.1519    -0.1519
#> rs115165870     -0.0358    -0.0358     0.1023    -0.0740     -0.0740    -0.0740
#> rs12727865      -0.1466    -0.1466     0.1971    -0.1472     -0.1472    -0.1472
#> rs147889978     -0.0162    -0.0162     0.0704    -0.0137     -0.0137    -0.0137
#> rs6702754       -0.1466    -0.1466     0.1971    -0.1472     -0.1472    -0.1472
#> rs1209264       -0.1308    -0.1308     0.3119    -0.1281     -0.1281    -0.1281
#>             rs61811362 rs12408461 rs55908418 rs73023348 rs73023349 rs72999415
#> rs6691345      -0.1218    -0.1537    -0.0547    -0.1519    -0.1528    -0.1537
#> rs115165870    -0.0471    -0.0748    -0.0198    -0.0740    -0.0609    -0.0748
#> rs12727865     -0.0976    -0.1453    -0.1330    -0.1472    -0.1445    -0.1453
#> rs147889978    -0.0735    -0.0145    -0.0357    -0.0137    -0.0144    -0.0145
#> rs6702754      -0.0976    -0.1453    -0.1330    -0.1472    -0.1445    -0.1453
#> rs1209264      -0.2594    -0.1265     0.0983    -0.1281    -0.1258    -0.1265
#>             rs11265632 rs72999419 rs41313910 rs41310893 rs4845647 rs4845648
#> rs6691345       0.2293    -0.1519    -0.1537    -0.0873    0.2293   -0.1218
#> rs115165870     0.1023    -0.0740    -0.0748    -0.0350    0.1023   -0.0471
#> rs12727865      0.1971    -0.1472    -0.1504    -0.1214    0.1971   -0.0976
#> rs147889978     0.0704    -0.0137    -0.0145    -0.0658    0.0704   -0.0735
#> rs6702754       0.1971    -0.1472    -0.1504    -0.1214    0.1971   -0.0976
#> rs1209264       0.3119    -0.1281    -0.1265    -0.3772    0.3119   -0.2594
#>             rs72999422 rs76289529 rs12750774 rs3811448 rs6681207 rs3811449
#> rs6691345      -0.1519     0.2432    -0.0502   -0.1519   -0.1502   -0.1519
#> rs115165870    -0.0740     0.0625    -0.0473   -0.0740   -0.0731   -0.0740
#> rs12727865     -0.1472    -0.0626     0.1828   -0.1472   -0.1439   -0.1472
#> rs147889978    -0.0137    -0.0679    -0.0658   -0.0137   -0.0128   -0.0137
#> rs6702754      -0.1472    -0.0626     0.1828   -0.1472   -0.1439   -0.1472
#> rs1209264      -0.1281     0.0931     0.1643   -0.1281   -0.1298   -0.1281
#>             rs115697580 rs61559765 rs60092776 rs4382717 rs60931472 rs41269915
#> rs6691345       -0.0646    -0.1519    -0.1519    0.2249    -0.1464    -0.1464
#> rs115165870     -0.0428    -0.0740    -0.0740    0.1032    -0.0748    -0.0748
#> rs12727865      -0.0777    -0.1472    -0.1472    0.1958    -0.1453    -0.1453
#> rs147889978     -0.0421    -0.0137    -0.0137    0.0712    -0.0145    -0.0145
#> rs6702754       -0.0777    -0.1472    -0.1472    0.1958    -0.1453    -0.1453
#> rs1209264       -0.3496    -0.1281    -0.1281    0.3109    -0.1265    -0.1265
#>             rs11265634 rs72698179 rs9427082 rs12144146 rs41302545 rs7539745
#> rs6691345       0.1291    -0.0519   -0.1218    -0.0458    -0.1218    0.1275
#> rs115165870    -0.0016    -0.0110   -0.0471    -0.0454    -0.0471    0.0499
#> rs12727865      0.0915    -0.1372   -0.0976     0.1897    -0.0976    0.0909
#> rs147889978     0.0691     0.0228   -0.0735    -0.0642    -0.0735    0.0762
#> rs6702754       0.0915    -0.1372   -0.0976     0.1897    -0.0976    0.0909
#> rs1209264       0.2733     0.0857   -0.2594     0.1618    -0.2594    0.2617
#>             rs7543174 rs56047170 rs7550664 rs7553602 rs4074436 rs4076089
#> rs6691345      0.1244    -0.1155   -0.1199   -0.1117    0.0708    0.0330
#> rs115165870   -0.0045    -0.0396   -0.0462   -0.0308   -0.0350   -0.1629
#> rs12727865     0.0929    -0.1025   -0.0998   -0.1057    0.1349   -0.0143
#> rs147889978    0.0669    -0.0831   -0.0727   -0.0727    0.0337   -0.0104
#> rs6702754      0.0929    -0.1025   -0.0998   -0.1057    0.1349   -0.0143
#> rs1209264      0.2643    -0.2672   -0.2615   -0.2615   -0.0571    0.0647
#>             rs4845649 rs72698182 rs149551556 rs9427089 rs56670270 rs4845651
#> rs6691345      0.0708    -0.0648     -0.0194    0.0330    -0.0433    0.0526
#> rs115165870   -0.0350    -0.0198     -0.0358   -0.1629     0.1459   -0.1316
#> rs12727865     0.1349    -0.1402      0.0000   -0.0143     0.0448   -0.0547
#> rs147889978    0.0337    -0.0357     -0.0352   -0.0104     0.0037    0.0022
#> rs6702754      0.1349    -0.1402      0.0000   -0.0143     0.0448   -0.0547
#> rs1209264     -0.0571     0.0983      0.0675    0.0647    -0.0759    0.0606
#>             rs6427741 rs150681049 rs6427743 rs61811370 rs61811371 rs4845652
#> rs6691345     -0.0638     -0.0611    0.0526    -0.0638    -0.0638   -0.0616
#> rs115165870   -0.0243     -0.0516   -0.1316    -0.0243    -0.0243   -0.0232
#> rs12727865    -0.1352     -0.0503   -0.0547    -0.1352    -0.1352   -0.1382
#> rs147889978   -0.0399     -0.0074    0.0022    -0.0399    -0.0399   -0.0388
#> rs6702754     -0.1352     -0.0503   -0.0547    -0.1352    -0.1352   -0.1382
#> rs1209264      0.0850      0.0251    0.0606     0.0850     0.0850    0.0830
#>             rs6660775 rs67860750 rs78038982 rs71651696 rs147483024 rs8192484
#> rs6691345     -0.0638    -0.0190    -0.0611     0.0193     -0.0799   -0.0493
#> rs115165870   -0.0243    -0.0047    -0.0516    -0.0627     -0.0213   -0.0250
#> rs12727865    -0.1352     0.1941    -0.0503     0.0288     -0.0398    0.0242
#> rs147889978   -0.0399    -0.0708    -0.0074    -0.0099     -0.0563   -0.0235
#> rs6702754     -0.1352     0.1941    -0.0503     0.0288     -0.0398    0.0242
#> rs1209264      0.0850     0.0554     0.0251     0.0099     -0.3589    0.0735
#>             rs61811372 rs4845378 rs78794625 rs12072348 rs2072659 rs2072660
#> rs6691345      -0.0638   -0.0544    -0.0433    -0.0682   -0.0736    0.1160
#> rs115165870    -0.0243   -0.0245     0.0980     0.0062    0.1838   -0.0918
#> rs12727865     -0.1423   -0.1294     0.0266    -0.0664   -0.0066   -0.2490
#> rs147889978    -0.0399   -0.0403     0.0037    -0.0239    0.1906   -0.1288
#> rs6702754      -0.1423   -0.1294     0.0266    -0.0664   -0.0066   -0.2490
#> rs1209264       0.0850    0.0859    -0.1029     0.1286    0.0512   -0.0934
#>             rs2072661 rs4292956 rs45490696 rs12130403 rs139826441 rs6680410
#> rs6691345     -0.0901   -0.0220    -0.0316    -0.0379     -0.0383   -0.0250
#> rs115165870    0.0808    0.2501     0.2531     0.0881      0.1045    0.2471
#> rs12727865     0.2399    0.0780     0.0842     0.1326      0.0584    0.0719
#> rs147889978    0.1312    0.3046     0.3079     0.0850      0.0069    0.3013
#> rs6702754      0.2399    0.0780     0.0842     0.1326      0.0584    0.0719
#> rs1209264      0.0946    0.0854     0.0831     0.0674     -0.0565    0.0877
#>             rs3811450 rs4845653
#> rs6691345     -0.0220   -0.0869
#> rs115165870    0.2501   -0.0028
#> rs12727865     0.0691   -0.1592
#> rs147889978    0.3046    0.0180
#> rs6702754      0.0691   -0.1592
#> rs1209264      0.0854    0.1147