GENI plots test data for the Manhattan plot

An R object containing formatted results from a genome-wide association study (GWAS) of interleukin-6 levels with p-values < 1e-3 (GCST90274815).

Usage

geni_test_manhattan

Format

A data.frame containing the variables necessary for constructing a static Manhattan plot. Specifically, this includes the following information:

chr the chromosome for each genomic marker

pos the genomic position (GRCh37) for each genomic marker

pvalue the association p-value for each genomic marker

highlight an indicator variable whether the genomic marker should be highlighted: 0 = point not highlighted, 1 = first highlight colour

highlight_shape an indicator variable of the illustrated shape for the genomic markers: 0 = standard circle, 5 = standard diamond with border

label the name of the nearest gene for each sentinel genome-wide associated genetic variant, the labels for non-sentinel genetic variants are omitted

Author

James Staley jrstaley95@gmail.com

Wes Spiller

Examples

head(geni_test_manhattan)
#>   chr     pos   pvalue highlight highlight_shape label
#> 1   1  963661 1.54e-04         0               0      
#> 2   1  968251 5.38e-04         0               0      
#> 3   1 1010040 1.67e-04         0               0      
#> 4   1 1112698 6.82e-04         0               0      
#> 5   1 2031103 8.97e-05         0               0      
#> 6   1 2036650 2.07e-04         0               0