GENI plots test data for the PheWAS plot

An R object containing formatted data from a phenome-wide association study (PheWAS) of rs2228145 (1:154426970-A-C) from FinnGen r9.

Usage

geni_test_phewas

Format

A data.frame containing the variables necessary for constructing an interactive PheWAS plot. Specifically, this includes the following information:

pvalue the association p-value for each phenotype

sign the direction of the association with the phenotype, where 0 = missing, 1 positive association, -1 negative association

group the phenotype group for each phenotype

label the phenotype name

text the hover text containing further information on the association including: phenotype, phenotypic category, genetic variant ID, direction of association, association p-value, number of cases and number of controls

Author

James Staley jrstaley95@gmail.com

Wes Spiller

Examples

head(geni_test_phewas)
#>   pvalue sign                    group
#> 1 0.3750   -1 AB1: Infectious diseases
#> 2 0.0506    1 AB1: Infectious diseases
#> 3 0.1120    1 AB1: Infectious diseases
#> 4 0.2870   -1 AB1: Infectious diseases
#> 5 0.0352   -1 AB1: Infectious diseases
#> 6 0.0113    1 AB1: Infectious diseases
#>                                                 label
#> 1                                       Actinomycosis
#> 2                                          Amoebiasis
#> 3   Anogenital herpesviral [herpes simplex] infection
#> 4                                       Aspergillosis
#> 5 Atypical virus infections of central nervous system
#> 6        Bacterial, viral and other infectious agents
#>                                                                                                                                                                                                  text
#> 1                                         Phenotype: Actinomycosis<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: -1<br>p-value: 0.375<br>N cases: 93<br>N controls: 332343
#> 2                                           Phenotype: Amoebiasis<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: 1<br>p-value: 0.0506<br>N cases: 149<br>N controls: 336396
#> 3    Phenotype: Anogenital herpesviral [herpes simplex] infection<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: 1<br>p-value: 0.112<br>N cases: 1812<br>N controls: 366340
#> 4                                        Phenotype: Aspergillosis<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: -1<br>p-value: 0.287<br>N cases: 194<br>N controls: 369098
#> 5 Phenotype: Atypical virus infections of central nervous system<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: -1<br>p-value: 0.0352<br>N cases: 260<br>N controls: 375157
#> 6        Phenotype: Bacterial, viral and other infectious agents<br>Category: AB1: Infectious diseases<br>rsID: rs2228145-C<br>Direction: 1<br>p-value: 0.0113<br>N cases: 2591<br>N controls: 374686